requirements.txt
selenium
webdriver-manager
pytest
allure-pytest
cryptography
pandas
openpyxl